+++ tests/Header/AcceptCharsetValueTest.php
@@ -14,7 +14,7 @@ class AcceptCharsetValueTest extends TestCase
 {
     public function testInterface()
     {
-        $a = Charset::maybe('unicode-1-1', $q = new Quality(0.8))->match(
+        $a = Charset::maybe('unicode-1-1', $q = Quality::of(80))->match(
             static fn($charset) => $charset,
             static fn() => null,
         );
@@ -23,23 +23,23 @@ class AcceptCharsetValueTest extends TestCase
         $this->assertSame($q, $a->quality());
         $this->assertSame('unicode-1-1;q=0.8', $a->toString());
 
-        Charset::maybe('iso-8859-5', new Quality(1))->match(
+        Charset::maybe('iso-8859-5', Quality::max())->match(
             static fn($charset) => $charset,
             static fn() => throw new \Exception,
         );
-        Charset::maybe('Shift_JIS', new Quality(1))->match(
+        Charset::maybe('Shift_JIS', Quality::max())->match(
             static fn($charset) => $charset,
             static fn() => throw new \Exception,
         );
-        Charset::maybe('ISO_8859-9:1989', new Quality(1))->match(
+        Charset::maybe('ISO_8859-9:1989', Quality::max())->match(
             static fn($charset) => $charset,
             static fn() => throw new \Exception,
         );
-        Charset::maybe('NF_Z_62-010_(1973)', new Quality(1))->match(
+        Charset::maybe('NF_Z_62-010_(1973)', Quality::max())->match(
             static fn($charset) => $charset,
             static fn() => throw new \Exception,
         );
-        Charset::maybe('*', new Quality(1))->match(
+        Charset::maybe('*', Quality::max())->match(
             static fn($charset) => $charset,
             static fn() => throw new \Exception,
         );
@@ -55,6 +55,7 @@ class AcceptCharsetValueTest extends TestCase
                     static fn() => null,
                 )
                 ?->quality()
+                ?->toParameter()
                 ?->value(),
         );
     }
@@ -62,7 +63,7 @@ class AcceptCharsetValueTest extends TestCase
     #[DataProvider('invalids')]
     public function testReturnNothingWhenInvalidAcceptCharsetValue($value)
     {
-        $this->assertNull(Charset::maybe($value, new Quality(1))->match(
+        $this->assertNull(Charset::maybe($value, Quality::max())->match(
             static fn($charset) => $charset,
             static fn() => null,
         ));